+++ gui/CleanCycleGUI/app/src/main/java/uk/ac/cam/cl/cleancyclegui/RouteHandler.java
@@ -1,0 +1,9 @@
+package uk.ac.cam.cl.cleancyclegui;
+
+import java.util.List;
+
+import uk.ac.cam.cl.cleancyclegraph.EdgeComplete;
+
+public interface RouteHandler {
+    void handleRoute(List<EdgeComplete> route);
+}
